backend/requirements.txt
bidict==0.23.1
cffi==2.0.0
dnspython==2.8.0
eventlet==0.40.4
greenlet==3.3.0
h11==0.16.0
numpy==2.4.0
pedalboard==0.9.19
pycparser==2.23
python-engineio==4.13.0
python-rtmidi==1.5.8
python-socketio==5.16.0
simple-websocket==1.1.0
sounddevice==0.5.3
wsproto==1.3.2
